Abandoning webgl rendering for 2d stuff:
It turns out the matrix math involved in sending transformation matrices to webgl outweigh the benefits of webgl rendering. The current method of rendering 2d things (applying translation, rotation and scale to the CanvasRenderingContext2D) is much faster even for tiny display lists. It may be that updating uniforms in a special '2d context' shader could be a way to get some extra speed but I'll leave that to interested parties later...